Add LazyJust strategy (and move Just to same file)
